started Assignment 3

+++ VetShop/src/main/java/com/veterinary/services/implementations/RegularUserServiceImpl.java
@@ -6,6 +6,8 @@ import com.veterinary.entities.Animal;
 import com.veterinary.entities.RegularUser;
 import com.veterinary.repositories.RegularUserRepository;
 import com.veterinary.services.RegularUserService;
+import com.veterinary.services.exceptions.AlreadyExistingException;
+import com.veterinary.services.exceptions.NoSuchEntityException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -16,21 +18,30 @@ import java.util.stream.Collectors;
 @Service
 public class RegularUserServiceImpl implements RegularUserService {
 
-    @Autowired
     private RegularUserRepository regUserRepo;
 
+    @Autowired
+    public RegularUserServiceImpl(RegularUserRepository regularUserRepository){
+        regUserRepo = regularUserRepository;
+    }
+
     @Override
-    public RegularUserDTO getByUsername(String username) {
-        return new RegularUserDTO(regUserRepo.findByUsername(username));
+    public RegularUserDTO getByUsername(String username) throws NoSuchEntityException {
+        RegularUser found = regUserRepo.findByUsername(username);
+        if(found == null)
+            throw new NoSuchEntityException("Invalid username");
+        return new RegularUserDTO(found);
     }
 
     @Override
     public List<RegularUserDTO> getAll() {
-        return  regUserRepo.findAll().stream().map(RegularUserDTO::new).collect(Collectors.toList());
+        return regUserRepo.findAll().stream().map(RegularUserDTO::new).collect(Collectors.toList());
     }
 
     @Override
-    public RegularUserDTO save(String username, String password, String fullName, String userType) {
+    public RegularUserDTO save(String username, String password, String fullName, String userType) throws AlreadyExistingException, NoSuchEntityException {
+        if(regUserRepo.findByUsername(username)!=null)
+            throw new AlreadyExistingException("Username taken");
         RegularUser regularUser = RegularUser.builder().consultations(new ArrayList<>()).build();
         regularUser.setFullName(fullName);
         regularUser.setPassword(password);
@@ -39,8 +50,10 @@ public class RegularUserServiceImpl implements RegularUserService {
     }
 
     @Override
-    public RegularUserDTO update(int id, String username, String password, String fullName) {
+    public RegularUserDTO update(int id, String username, String password, String fullName) throws NoSuchEntityException {
         RegularUser regularUser = regUserRepo.findById(id).orElse(null);
+        if(regularUser == null)
+            throw new NoSuchEntityException("No such username");
         regularUser.setUsername(username);
         regularUser.setPassword(password);
         regularUser.setFullName(fullName);
@@ -49,8 +62,10 @@ public class RegularUserServiceImpl implements RegularUserService {
     }
 
     @Override
-    public RegularUserDTO delete(int id) {
+    public RegularUserDTO delete(int id) throws NoSuchEntityException {
         RegularUser regularUser = regUserRepo.findById(id).orElse(null);
+        if(regularUser == null)
+            throw new NoSuchEntityException("No such username");
         regUserRepo.delete(regularUser);
         return new RegularUserDTO(regularUser);
     }
